# api/docs/openapi-full.yaml
openapi: 3.0.3
info:
  title: GlobalReach API
  description: Enterprise Email Marketing Platform API
  version: 2.0.0
  contact:
    email: support@globalreach.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://api.globalreach.com/api/v1
    description: Production server
  - url: https://staging.api.globalreach.com/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: AUTHENTICATION_FAILED
        message:
          type: string
          example: Invalid credentials
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
...
    EmailAccount:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        email:
          type: string
        platform:
          type: string
          enum: [gmail, outlook, sendgrid, custom]
        status:
          type: string
          enum: [active, inactive, verified]
        createdAt:
          type: string
          format: date-time
    Campaign:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [marketing, transactional, newsletter]
        status:
          type: string
          enum: [draft, active, completed, paused]
        emailCount:
          type: integer
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /accounts:
    get:
      tags: [Accounts]
      summary: Get user email accounts
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmailAccount'
    post:
      tags: [Accounts]
      summary: Create email account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                platform:
                  type: string
                smtpHost:
                  type: string
                smtpPort:
                  type: integer
                smtpUser:
                  type: string
                smtpPass:
                  type: string
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/EmailAccount'
  /accounts/{id}:
    get:
      tags: [Accounts]
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/EmailAccount'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags: [Accounts]
      summary: Update account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                platform:
                  type: string
      responses:
        '200':
          description: Account updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
    delete:
      tags: [Accounts]
      summary: Delete account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /campaigns:
    get:
      tags: [Campaigns]
      summary: Get user campaigns
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: List of campaigns
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/Campaign'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                          limit:
                            type: integer
                          total:
                            type: integer
                          pages:
                            type: integer
    post:
      tags: [Campaigns]
      summary: Create campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                type:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Campaign created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Campaign'
...